Refac(new Asset)
Squashed commit of the following:

commit 10f9f7d9045efce17148d96cb861c456daf3795b
Author: odoldotol <odoldotol@gmail.com>
Date:   Sun Aug 20 23:36:35 2023 +0900

    Remove(기존 코드), Rename(- => _)

commit 8deffec1e675e7a0eed7f438a58ea756f604542b
Author: odoldotol <odoldotol@gmail.com>
Date:   Sun Aug 20 23:02:12 2023 +0900

    기존 코드 제거

commit 4e6aed863cb29f896428d103963c7f42b487607f
Author: odoldotol <odoldotol@gmail.com>
Date:   Sun Aug 20 22:45:44 2023 +0900

    기존 코드 그대로 이식

commit e626cd31dd60a447bf3b001fb357a3932b8b1d9f
Author: odoldotol <odoldotol@gmail.com>
Date:   Sun Aug 20 21:39:12 2023 +0900

    init
