+++ market/src/asset/decorator/index.ts
@@ -1,0 +1,3 @@
+export * from './api_addAssets.decorator';
+export * from './api_getPriceByTicker.decorator';
+export * from './api_getPriceByExchange.decorator';
