+++ market/src/asset/response/getPriceByTicker.response.ts
