+++ market/src/dev/dev.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
 import { DevController } from './dev.controller';
 import { DevService } from './dev.service';
-import { DBModule } from 'src/database/database.module';
+import { DatabaseModule } from 'src/database/database.module';
 import { MarketModule } from 'src/market/market.module';
 
 @Module({
   imports: [
-    DBModule,
+    DatabaseModule,
     MarketModule
   ],
   controllers: [DevController],
