+++ market/src/market/child_api/child_api.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { HttpModule } from '@nestjs/axios';
 import { ConnectionService } from './connection.service';
-import { ChildApiService } from './child-api.service';
+import { ChildApiService } from './child_api.service';
 import { EnvKey } from 'src/common/enum/envKey.emun';
 import { EnvironmentVariables } from 'src/common/interface/environmentVariables.interface';
