+++ market/src/market/exchange.service.ts
@@ -1,6 +1,6 @@
 import { Inject, Injectable, Logger, OnModuleInit } from "@nestjs/common";
 import { TExchangeCore } from "src/common/type/exchange.type";
-import { ChildApiService } from "./child-api/child-api.service";
+import { ChildApiService } from "./child_api/child_api.service";
 import { Exchange } from "./class/exchange";
 import { EMarketEvent } from "./enum/eventName.enum";
 import { ExchangeContainer } from "./exchangeContainer";
