+++ market/src/market/market.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { MarketService } from './market.service';
-import { ChildApiModule } from './child-api/child-api.module';
+import { ChildApiModule } from './child_api/child_api.module';
 import { ExchangeService } from './exchange.service';
 import { ExchangeContainer } from './exchangeContainer';
 import { ExchangeConfigArrProvider } from './provider/exchangeConfigArr.provider';
