+++ market/src/market/mock/childApiService.mock.ts
@@ -1,4 +1,4 @@
-import { ChildApiService } from "../child-api/child-api.service";
+import { ChildApiService } from "../child_api/child_api.service";
 
 export const mockChildApiService = {
 } as unknown as ChildApiService;
