+++ market/src/updater/updater.module.ts
@@ -2,13 +2,13 @@ import { Module } from '@nestjs/common';
 import { UpdaterController } from './updater.controller';
 import { UpdaterService } from './updater.service';
 import { MarketModule } from 'src/market/market.module';
-import { DBModule } from 'src/database/database.module';
-import { ProductApiModule } from 'src/product-api/product-api.module';
+import { DatabaseModule } from 'src/database/database.module';
+import { ProductApiModule } from 'src/product_api/product_api.module';
 
 @Module({
   imports: [
     MarketModule,
-    DBModule,
+    DatabaseModule,
     ProductApiModule
   ],
   controllers: [UpdaterController],
