+++ market/src/updater/updater.service.ts
@@ -4,9 +4,7 @@ import { SchedulerRegistry } from '@nestjs/schedule';
 import { CronJob } from 'cron';
 import { MarketService } from 'src/market/market.service';
 import { DBRepository } from 'src/database/database.repository';
-import { ProductApiService } from 'src/product-api/product-api.service';
-import { AddAssetsResponse } from './response/addAssets.response';
-import { Either } from "src/common/class/either";
+import { ProductApiService } from 'src/product_api/product_api.service';
 import { pipe, map, toArray, toAsync, tap, each, filter, concurrent, curry } from "@fxts/core"; // Todo: 제거
 import * as F from "@fxts/core";
 import { EnvironmentVariables } from 'src/common/interface/environmentVariables.interface';
@@ -17,7 +15,6 @@ import { ExchangeService as DbExchangeService } from 'src/database/exchange/exch
 import { FinancialAssetService as DbFinancialAssetService } from 'src/database/financialAsset/financialAsset.service';
 import { Yf_infoService as DbYfInfoService } from 'src/database/yf_info/yf_info.service';
 import { Exchange } from 'src/market/class/exchange';
-import { Exchange as ExchangeEntity } from 'src/database/exchange/exchange.entity';
 
 /**
  * ### TODO: Refac:
@@ -84,7 +81,7 @@ export class UpdaterService implements OnModuleInit {
   }
 
   // Todo: Refac - Exchange 리팩터링 후 억지로 끼워맞춤
-  private async updateAssetsOfExchange(exchange: Exchange, launcher: LogPriceUpdate["launcher"]) {
+  public async updateAssetsOfExchange(exchange: Exchange, launcher: LogPriceUpdate["launcher"]) {
     const ISO_Code = exchange.ISO_Code;
     const yf_exchangeTimezoneName = exchange.ISO_TimezoneName;
     const previous_close = exchange.getMarketDate().toISOString();
@@ -154,95 +151,6 @@ export class UpdaterService implements OnModuleInit {
       };
     };
     rq();
-  };
-
-  // TODO: Refac - 기능 분리
-  // Todo: Refac - Exchange 리팩터링 후 아직 함께 이팩터링되지 못함
-  addAssets = async (tickerArr: string[]) => {
-    const response = new AddAssetsResponse();
-    const spMap: Map<string, string[]> = new Map();
-
-    await pipe( // 중복제거와 exists필터 부분은 단일 티커처리시 필요없음. 이 부분 보완하기
-      new Set(tickerArr).values(), toAsync,
-      map(this.eitherFilter_existsAsset),
-      map(ele => ele.flatMap(this.marketService.fetchInfo.bind(this.marketService))),
-      map(ele => ele.flatMap(this.fulfillYfInfo.bind(this))),
-      filter(ele => ele.isLeft() ? (response.failure.info.push(ele.getLeft), false) : true),
-      map(ele => ele.getRight),
-      concurrent(this.CHILD_CONCURRENCY),
-      toArray,
-      tap(arr => this.dbYfInfoSrv.insertMany(arr)
-        .then(res => response.success.info = res)
-        .catch(err => (response.failure.info = response.failure.info.concat(err.writeErrors),
-          response.success.info = response.success.info.concat(err.insertedDocs)))),
-      each(ele => spMap.has(ele.exchangeTimezoneName) ?
-        spMap.get(ele.exchangeTimezoneName)?.push(ele.symbol) // ?
-        : spMap.set(ele.exchangeTimezoneName, [ele.symbol])));
-    await pipe(
-      spMap, toAsync,
-      filter(this.isNewExchange.bind(this)),
-      map(this.applyNewExchange),
-      each(ele => ele.isRight() ?
-      // @ts-ignore // exchange 리팩터링 후 문제
-        response.success.status_price.push(ele.getRight)
-        : response.failure.status_price.push(ele.getLeft)));
-    return response;
-  };
-
-  private eitherFilter_existsAsset = async (ticker: string): Promise<Either<any, string>> =>
-    (await this.dbYfInfoSrv.exists(ticker) === null) ?
-      Either.right(ticker) : Either.left({ msg: "Already exists", ticker });
-
-  // Todo: Refac - Exchange 리팩터링 후 억지로 끼워맞춤
-  private fulfillYfInfo = async (info: YfInfo): Promise<Either<any, FulfilledYfInfo>> => {
-    const exchange = this.mkExchangeSrv.findExchange(info.exchangeTimezoneName)! //
-    const ISO_Code = exchange.ISO_Code; //
-    return ISO_Code === undefined ?
-      Either.left({
-        msg: "Could not find ISO_Code",
-        yf_exchangeTimezoneName: info.exchangeTimezoneName,
-        yfSymbol: info.symbol
-      })
-      : Either.right({
-        ...info,
-        regularMarketLastClose: !exchange.isMarketOpen() ? info.regularMarketPrice : info.regularMarketPreviousClose
-      })
-  };
-
-  private async isNewExchange([yf_exchangeTimezoneName, _]: [string, string[]]) {
-    return !(await this.dbExchangeSrv.exist({ ISO_TimezoneName: yf_exchangeTimezoneName }));
-  }
-
-  // TODO - Refac
-  // Todo: Refac - Exchange 리팩터링 후 억지로 끼워맞춤
-  private applyNewExchange = async ([yf_exchangeTimezoneName, symbolArr]: [string, string[]]): Promise<Either<any, ExchangeEntity>> => {
-    const exchange = this.mkExchangeSrv.findExchange(yf_exchangeTimezoneName)! //
-    const ISO_Code = exchange.ISO_Code; //
-    if (ISO_Code === undefined) {
-      this.logger.error(`${symbolArr[0]} : Could not find ISO_Code for ${yf_exchangeTimezoneName}`);
-      return Either.left({
-        msg: "Could not find ISO_Code",
-        yf_exchangeTimezoneName,
-        symbol: symbolArr
-      });
-    } else {
-      return await this.dbExchangeSrv.createOne({
-        ISO_Code,
-        ISO_TimezoneName: yf_exchangeTimezoneName,
-        marketDate: exchange.getMarketDate().toISOString()
-      }).then(async res => {
-        this.logger.verbose(`${ISO_Code} : Created new Exchange`);
-        this.mkExchangeSrv.registerUpdater(this.updateAssetsOfExchange.bind(this), res);
-        return Either.right(res);
-      }).catch(error => {
-        this.logger.error(error);
-        return Either.left({
-          error,
-          yf_exchangeTimezoneName,
-          symbol: symbolArr
-        });
-      });
-    };
   };
 
   // @ts-ignore
